+++ setup.py
@@ -86,23 +86,22 @@ if os.getenv('MYPY_USE_MYPYC', None) == '1':
 
 if USE_MYPYC:
     MYPYC_BLACKLIST = (
-        # Designed to collect things that can't be compiled
-        'mypyc_hacks.py',
-        'interpreted_plugin.py',
-
         # Need to be runnable as scripts
         '__main__.py',
         'sitepkgs.py',
         'dmypy.py',
 
-        # We lie to mypy about code reachability here
+        # Needs to be interpreted to provide a hook to interpreted plugins
+        'interpreted_plugin.py',
+
+        # Uses __getattr__/__setattr__
+        'split_namespace.py',
+
+        # Lies to mypy about code reachability
         'bogus_type.py',
 
-        # Something goes wrong
-        'main.py',
-
         # We don't populate __file__ properly at the top level or something?
-        # Also I think there would be problems with how we generate new version.pys
+        # Also I think there would be problems with how we generate version.py.
         'version.py',
     )
